Cin 62 back 리뷰수정,삭제 api

+++ src/main/java/com/back/cinetalk/movie/dto/MovieDTO.java
@@ -18,7 +18,7 @@ public class MovieDTO {
 
     private Long id;
 
-    private Long movieId;
+    private int movieId;
 
     private String movienm;
